storage-setup: Setup LVM thin pool and modify docker-storage file accordingly
If users passes "SETUP_LVM_THIN_POOL=yes" in docker-storage-setup config
file, then setup an LVM thin pool and pass it to docker. That way docker
does not have to setup a pool and it can just concern itself with creation
and deletion of thin and snap devices.

This can allow lvm to repair pool and autoamtically grow pool when less
space is left in pool.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>