storage-setup: Check for volume attributes to figure out if thin pool exists
lvm_pool_exists() checks whether lvm thin pool volume already exists or
not. It comapres existing volume names with pool name "docker-pool".

Problem is that name "docker-pool" is used for data volume too and if
lvm_pool_exists() is called after data volume creation, then it thinks
that pool already exists. But the fact is that docker-pool is data volume
and not the pool volume.

Fix this problem by also checking volume attributes. A pool volume will
start with "t".

A little background on why data volume name is same as pool volume name.
lvconvert takes the data volume name and uses it as pool volume name.
So data volume name was kept as dokcer-pool so that finally it ends up as
pool name. This probably can be managed better by doing "lvrename" on pool
after creation, but that's a todo improvement.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>