d-s-s: Remove support for passing data and meta block devices to docker
Docker does not do all the error handing which can be required if thin
pool gets full or thin pool runs into issues. So passing block devices
to docker and let it setup thin pool is not a very good idea from error
handling point of view.

Instead always setup lvm thin pool and pass that to docker so that docker
does not have to take care of pool management issues and lvm2 commands
can take care of it. (See man lvmthin).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>