docker-storage-setup: Use growpart only if it is available
Trying to make docker-storage-setup more generic so that it can be shipped
with non atomic and non-cloud images too. There we might not have "growpart"
utility available. So remove dependency on growpart. Use it only when
available. Where we need it out of box, we need to make sure that
cloud-utils-growpart gets installed.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>