docker-storage-setup: Create helper functions to create partitions and volume group
If user passes in DEVS= option, we do various checks and create partitions
on disks. And then we either create a new volume group or add disks to the
specified volume group.

Create two functions. One for creating partitions and other for volume
group handling. And move relevant code in those functions. This should not
be any functionality change. Just code re-organization.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>