docker-storage-setup: Add MIN_DATA_SIZE configuration knob
Add a configuratoin knob MIN_DATA_SIZE which allows users to specify minimum
size of data volume (or pool volume). This is needed as in many a cases our
pool is very small to begin with and even first image pull into the pool hangs.

Keep minimum pool size to be 2G as default. This can be changed as needed. 

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>