docker-storage-setup: Relax restriction on data size suffixes
Previously we were not doing any checks on DATA_SIZE suffixes and passing
everything to lvm directly. It turns out lvm just checks if things are
starting with valid suffix and anythign after that is acceptable. For
example -L 5GHello is acceptable and lvm will treat it as 5GiB.

Then with MIN_DATA_SIZE patch, we put limitations on suffixes. One could
use only one valid suffix character and nothing after that. For example,
only 5G is acceptable and 5GB and 5GiB are failing.

This is being seen as regression.

This patch allows now [bB] and i[bB] after the valid suffix. That is,
values of type 5G, 5GB, 5GiB are acceptable and all will be treated
as 5GiB.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>