docker-storage-setup: Exit early if a valid volume group is not specified/found
Right now docker-storage-setup assumes that root device is a lvm logical
volume and there is a valid volume group backing it. But if user is not
using lvm and root devices is just a block device or a partition, then
this assumption is not valid any more.

Detect such cases and error out early gracefully.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>