Use a different method to query client version information
Looks like docker changed the format of "docker version" output and current
code is broken. We are grepping for "Client version" and that string is no
more there.

Now only query for client version using "docker version --format" option.
This method works only docker 1.8 onwards. That means if this script is
used with docker 1.7 or older, it will not get version information.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>