Suppress lvm error messages when probing for volume group backing rootfs
When we are probing for volume group backing rootfs, then lvm will give
out error messages if rootfs device is not a logical volume. For example.

  Volume group "vda1" not found
  Cannot process volume group vda1

These messages are confusing and are not useful for the user. And this
easily happens if docker-storage-setup runs on a system where lvm is
not being used. 

Supress these messages. Instead just output an INFO message saying rootfs
volume group could not be determined.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>