Do not exit early if rootfs is not on a logical volume
Right now if rootfs is not backed by a volume group and user has not specified
a volume group we exit early saying a valid VG is required. This was true
as long as we supported only devicemapper driver. But now we support overlay
driver too and we don't have to have a valid volume group for that. 

So instead of exiting early, move this check later when we know we are
dealing with devicemapper storage driver.

Also along the way we need to make sure various operations like creating
disk partitions, growing pvs and growing rootfs are not done if we don't
have rootfs on a logical volume.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>