Source local libdss.sh first
First see if a libdss.sh is present in same direcotry as docker-storage-setup
and source that first. If not, then source the one present in standard
location. This helps with doing development where one can modify libdss.sh
in git tree and expects docker-storage-setup.sh to work and be able to see
changes of libdss.sh.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>