Do not create a new thin pool if docker is already having devicemapper metadata
If docker already has devicemapper metadata dir, do not create a new thin
pool. This most likely means that docker was already started and it used
some other thin pool (most likely using loop devices) and now we are setting
up a new thin pool and asking docker to use it. This will not work. Old
metadata of docker will fail to work with new thin pool.

This is a common situation and lot of people have complained about running
into this. Right now docker message is  that base device UUID verification
failed but that does not tell user much about what to do.

So detect this in docker-storage-setup and error out and display a message
which explains how one can do complete cleanup. This will destroy all 
existing images and containers and all docker metadata.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #117
Approved by: rhvgoyal