Output more info about waiting for device
I am dealing with issues where it looks like that device node did not
become available before pvcreate/wipefs was run on it and it failed.
Output more info messages which can give out more information if we
waited for device node or not and whether device node became available
or not.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #125
Approved by: rhatdan