Convert device names to abosolute path under /dev/
Right now we are allowing DEVS to be specified just by their name say "sdb"
and in the code we assume it is /dev/sdb. This is sprinkled all over the
code and is easy to break.

Recenlty I noticed that logic to detect if a partition is already part of
volume group does not work if DEVS=vdb is specified. But same logic works
fir DEVS=/dev/vdb is specified.

pv_name=$(pvs --noheadings -o pv_name -S pv_name=$dev,vg_name=$vg

Here pvname needs to be /dev/vdb1 and just specifying vdb1 does not
return anything. And code assumes that disk is not part of volume
group. And tries to wipe signatures and create partitiona and wipe
signature fails as.

wipefs: error: /dev/vdb: probing initialization failed: Device or resource busy
ERROR: Failed to wipe signatures on device /dev/vdb

So bottom line, it can be confusing and it is easy to break code.

So convert all device names to absolute path starting with /dev/ in
the beginning of script and in rest of the code just assume device
names are of the form /dev/foo. And that reduces confusion and
fixes the bug I mentioned above.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #131
Approved by: cgwalters