tests: Use sfdisk if parted is not present
Right now atomic host does not seem to include parted. In that case use
sfdisk to remove partitions.

Also wait for udev to settle down after partition deletion. In some
cases it has been observed that partition table reload failed after
partition deletion as device was busy. Suspicion is that udev is
keeping the device busy. So wait for udev to settle down. Thanks
to Jonathan Lebon for the idea and fix.

Signed-off-by: Jonathan Lebon <jlebon@redhat.com>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #185
Approved by: jlebon