Stop using service file for mounting extra volumes
Have container-storage-setup mount it, itself.

We want to remove the generation of the mount.target unit file, since
it is potentially fragile.  If a user modifies the
/etc/sysconfig/docker-storage-setup, there is a potential for the script to
leave on of these mount target files behind. With this change the mount
point could be left mounted, but the next reboot would clean it up.
Now when docker-storage-setup is called, it will not only take care of creating
the new volume but also mounting it.

If the volume already exists, then docker-storage-setup itself will now mount
the volume, if it is already not mounted.

Note: we have to leave around the --reset function, since we can not figure out
which mount targets were created in the past.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>

Closes: #202
Approved by: rhvgoyal