Do not default to name docker-pool if not in docker compatibility mode
By default name docker-pool should be used only for compatibility mode.
In new mode, user need to specify name of thin pool.

We had committed some code where name of thin pool could be default
to docker-pool even in new mode. But that code was only in upstream
tree for a very short duration. We should not have done that.

But can't carry code to support backward compatibility for that. It
becomes too messy.

Now onwards backward compatibility will be supported for docker
compatibility mode and backward compatibility for rest of the
stuff will be provided once we reach version 1.0.0. Till then
things are in a state of flux as lot of new things are being
added.

Moved CONTAINER_THINPOOL check earlier to make sure in non-compat
mode we check early that a CONTAINER_THINPOOL name has been 
passed in.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #227
Approved by: rhatdan