Remove volume group and partitions from disk during reset
When a storage configuration is removed, also remove the volume group
and remove partitions from the disk. This is done only if
container-storage-setup created a volume group. If a volume group is
already found on the system at the time of storage creation, then it
is left untouched and no disks are removed as well.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #243
Approved by: rhatdan