Allow for extents based size for ROOT_SIZE
There are two ways to specify a size to LVM. One is -l (extents)
and one is -L (size). In absolute terms -L is nice, but in relative
terms (i.e. use a percentage of something) -l is nice. Allow for users
to provide a % based option with this commit.

man lvresize for more information.

Fixes #257

Closes: #258
Approved by: cgwalters