Check ftype=1 for xfs filesystem
If ftype=0 for underlying xfs (for overlay), then whiteouts become visible.
So make sure ftype=1 otherwise container-storage-setup fails.

This needs to be tested on xfs which will be used by container runtime
as underlying fs for overlayfs. We don't know about it for sure and
we are just guessing. This is hacky and hopefully down the line we can
pass this information using a config variable (if need be).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #263
Approved by: cgwalters