Give info if storage is already configured with new STORAGE_DRIVER
If storage is already configured, I think ideally we should not parse
/etc/syconfig/docker-storage-seutp anymore and exit. IOW, /etc/sysconfig-docker-storage-setup should be parsed only if a new storage configuration is
being setup.

That way for new settings to take affect, a user will have to first reset
storage and setup new storage.

But unfortunately that's not the case currently. For example, one can
enable/disable thin pool extension with currently setup storage just
by changing /etc/sysconfig/docker-storage-setup and restarting the
service.

Over upgrade we want to change default to overlay2 and if system is
already configured with devicemapper, it gives error message that
storage is already configured and can't switch to overlay2 and exits
with error code.

Change this to an Info message and exit with 0. That way we continue
to use currently setup storage and docker will not fail either. 

And I think this is probably right thing to do. Ideally, we should
not be parsing /etc/sysocnfig/docker-storage-setup if storage is
already setup. We can't enforce changes in
/etc/sysconfig/docker-storage-setup on existing storage configuration.
That's too hard to do.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>

Closes: #263
Approved by: cgwalters