During reset, first try to reset already configured thin pool
During reset we look at /etc/sysconfig/docker-storage-setup and figure
out which driver is in use and reset thin pool accordingly.

But now atomic upgrade seems to overwrite /etc/sysconfig/docker-storage-setup
and changes STORAGE_DRIVER=overlay2. That means reset path is broken.

If /etc/sysconfig/docker-storage is still around (and not corrupt), look
at it and find out configured thin pool and associated VG and reset
that. If that information is not present then look at
/etc/sysconfig/docker-storage-setup and continue as usual.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>